+++ module/Doctorant/src/Doctorant/Form/MissionEnseignement/MissionEnseignementHydratorFactory.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Doctorant\Form\MissionEnseignement;
+
+use Psr\Container\ContainerInterface;
+
+class MissionEnseignementHydratorFactory {
+
+    public function __invoke(ContainerInterface $container) : MissionEnseignementHydrator
+    {
+        $hydrator = new MissionEnseignementHydrator();
+        return $hydrator;
+    }
+}