+++ module/Doctorant/src/Doctorant/Provider/Privilege/MissionEnseignementPrivileges.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Doctorant\Provider\Privilege;
+
+use UnicaenAuth\Provider\Privilege\Privileges;
+
+class MissionEnseignementPrivileges extends Privileges
+{
+    const MISSION_ENSEIGNEMENT_VISUALISER          = 'missionenseignement-missionenseignement_visualiser';
+    const MISSION_ENSEIGNEMENT_MODIFIER            = 'missionenseignement-missionenseignement_modifier';
+}