+++ module/Doctorant/src/Doctorant/Service/MissionEnseignement/MissionEnseignementServiceAwareTrait.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Doctorant\Service\MissionEnseignement;
+
+trait MissionEnseignementServiceAwareTrait
+{
+    private MissionEnseignementService $missionEnseignementService;
+
+    public function getMissionEnseignementService(): MissionEnseignementService
+    {
+        return $this->missionEnseignementService;
+    }
+
+    public function setMissionEnseignementService(MissionEnseignementService $missionEnseignementService): void
+    {
+        $this->missionEnseignementService = $missionEnseignementService;
+    }
+
+
+}